shared function to avoid duplication of code
